+++ project1/src/constants/role.ts
@@ -1,0 +1,1 @@
+export const ADMIN_ROLE = "admin";
